wifibox-alpine: import Linux dev packages for building drivers
The `-dev` packages are needed for the custom drivers to build,
and it is better to import them so that the CI will not struggle
with building them every time.  Note that these packages are not
needed by the consumers, this is just a technical piece for
supporting development work.